* commit 'ebd494a557e70360534019691b32e0717f7986d4':
  JA-695

+++ app/src/main/java/life/plank/juna/zone/data/model/BoardEvent.kt
@@ -6,7 +6,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class BoardEvent(
         var type: String,
-        var foreignId: Int,
+        var foreignId: Long,
         var homeTeamLogo: String? = null,
         var awayTeamLogo: String? = null,
         var leagueName: String? = null
